Rename types/index to types/birdData
This will be a more suitable name, as in the upcoming changes, the
birdSlice reducer will be split into several, and therefore there
should be a type file for each.
