+++ app/javascript/react_app/components/bird/details_modal.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
 import Modal from '../shared/modal'
 
-import { BirdWithOrWithoutObservation, UserSettings } from '../../types'
+import { BirdWithOrWithoutObservation, UserSettings } from '../../types/birdData'
 import PopulationBars from './population_bars'
 import { populationInfo, migrationText } from '../../helpers/population_info'
 import BirdHouse from './bird_house'
