+++ app/javascript/react_app/components/bird/population_tooltip.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { BirdWithOrWithoutObservation } from '../../types'
+import { BirdWithOrWithoutObservation } from '../../types/birdData'
 import { isBreedingBird } from '../../helpers/breeding_bird'
 import PopulationBars from './population_bars'
 import { populationInfo } from '../../helpers/population_info'
