+++ app/javascript/react_app/types/birdData.ts
